# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Arcadia Backend",
        "title": "Arcadia API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/constants": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all Constants",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all Constants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.ConstantsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a constant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update a constant",
                "parameters": [
                    {
                        "enum": [
                            "minicons_in_lineup",
                            "matchmaking_rank_range",
                            "min_trophy_gain",
                            "trophy_gain_range",
                            "trophy_diff_loser",
                            "survivor_trophy_range",
                            "default_trophy_count",
                            "successive_duplicate_match_limit",
                            "daily_attack_limit",
                            "incr_xp_minicon",
                            "incr_xp_user",
                            "xp_level_multiplier",
                            "xp_base_count",
                            "level_up_stat_multiplier_numerator",
                            "level_up_stat_multiplier_denominator",
                            "is_arena_open",
                            "max_minicon_level",
                            "max_unlocked_minicons",
                            "type_multiplier_numerator",
                            "type_multiplier_denominator"
                        ],
                        "type": "string",
                        "description": "Name of the constant",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "New value of the constant",
                        "name": "newValue",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error in updating constants",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/lootbox": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all lootboxes of the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Get Lootboxes",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/controller.LootboxGETResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/lootbox/open": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Open a lootbox",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Open Lootbox",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player's Tile X coordinate",
                        "name": "x",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Player's Tile Y coordinate",
                        "name": "y",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Lootbox ID",
                        "name": "lootboxID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/controller.LootboxOpenPOSTResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Minicon Limit Reached",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Lootbox already opened",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "418": {
                        "description": "LOL, Nice Try",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.ConstantsResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "$ref": "#/definitions/model.ConstantType"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
...
        "controller.Lootbox": {
            "type": "object",
            "properties": {
                "isOpen": {
                    "description": "Is the lootbox open?",
                    "type": "boolean"
                },
                "lootboxID": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "x": {
                    "description": "Encrypted X coordinate of the lootbox",
                    "type": "string"
                },
                "y": {
                    "description": "Encrypted Y coordinate of the lootbox",
                    "type": "string"
                }
            }
        },
        "controller.LootboxGETResponse": {
            "type": "object",
            "properties": {
                "lootboxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Lootbox"
                    }
                }
            }
        },
        "controller.LootboxOpenPOSTResponse": {
            "type": "object",
            "properties": {
                "lootboxID": {
                    "type": "integer"
                },
                "unlocked": {
                    "$ref": "#/definitions/controller.Unlocked"
                }
            }
        },
...
        "controller.Unlocked": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "helper.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ConstantType": {
            "type": "string",
            "enum": [
                "minicons_in_lineup",
                "matchmaking_rank_range",
                "min_trophy_gain",
                "trophy_gain_range",
                "trophy_diff_loser",
                "survivor_trophy_range",
                "default_trophy_count",
                "successive_duplicate_match_limit",
                "daily_attack_limit",
                "incr_xp_minicon",
                "incr_xp_user",
                "xp_level_multiplier",
                "xp_base_count",
                "level_up_stat_multiplier_numerator",
                "level_up_stat_multiplier_denominator",
                "is_arena_open",
                "max_minicon_level",
                "max_unlocked_minicons",
                "type_multiplier_numerator",
                "type_multiplier_denominator"
            ],
            "x-enum-varnames": [
                "MiniconsInLineup",
                "MatchmakingRankRange",
                "MinTrophyGain",
                "TrophyGainRange",
                "TrophyDiffLoser",
                "SurvivorTrophyRange",
                "DefaultTrophyCount",
                "SuccessiveDuplicateMatchLimit",
                "DailyAttackLimit",
                "IncrXpMinicon",
                "IncrXpUser",
                "XpLevelMultiplier",
                "XpBaseCount",
                "LevelUpStatMultiplierNumerator",
                "LevelUpStatMultiplierDenominator",
                "IsArenaOpen",
                "MaxMiniconLevel",
                "MaxUnlockedMinicons",
                "TypeMultiplierNumerator",
                "TypeMultiplierDenominator"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}